+++ src/Oro/Bundle/UserBundle/Migrations/Schema/v1_0/OroUserBundle.php
@@ -108,7 +108,7 @@ class OroUserBundle implements Migration
             $table->addColumn('imap_configuration_id', 'integer', ['notnull' => false]);
         }
         $table->addColumn('business_unit_owner_id', 'integer', ['notnull' => false]);
-        $table->addColumn('status_id', 'smallint', ['notnull' => false]);
+        $table->addColumn('status_id', 'integer', ['notnull' => false]);
         $table->addColumn('username', 'string', ['length' => 255]);
         $table->addColumn('email', 'string', ['length' => 255]);
         $table->addColumn('name_prefix', 'string', ['notnull' => false, 'length' => 255]);
@@ -150,7 +150,7 @@ class OroUserBundle implements Migration
         /** Generate table oro_user_access_group **/
         $table = $schema->createTable('oro_user_access_group');
         $table->addColumn('user_id', 'integer', []);
-        $table->addColumn('group_id', 'smallint', []);
+        $table->addColumn('group_id', 'integer', []);
         $table->setPrimaryKey(['user_id', 'group_id']);
         $table->addIndex(['user_id'], 'IDX_EC003EF3A76ED395', []);
         $table->addIndex(['group_id'], 'IDX_EC003EF3FE54D947', []);
@@ -235,7 +235,7 @@ class OroUserBundle implements Migration
     {
         /** Generate table oro_user_email **/
         $table = $schema->createTable('oro_user_email');
-        $table->addColumn('id', 'smallint', ['autoincrement' => true]);
+        $table->addColumn('id', 'integer', ['autoincrement' => true]);
         $table->addColumn('user_id', 'integer', ['notnull' => false]);
         $table->addColumn('email', 'string', ['length' => 255]);
         $table->setPrimaryKey(['id']);
@@ -252,7 +252,7 @@ class OroUserBundle implements Migration
     {
         /** Generate table oro_user_status **/
         $table = $schema->createTable('oro_user_status');
-        $table->addColumn('id', 'smallint', ['autoincrement' => true]);
+        $table->addColumn('id', 'integer', ['autoincrement' => true]);
         $table->addColumn('user_id', 'integer', ['notnull' => false]);
         $table->addColumn('status', 'string', ['length' => 255]);
         $table->addColumn('created_at', 'datetime', []);
